moved files in created array package
